refactor : domain으로 일급 콜렉션 이동
